feat(Line): 지하철 노선 정보와 연결된 지하철 정보를 갖고 있는 Enum
